Refactor parameters handling: Move to a separate class
